requirements.txt
telethon
BeautifulSoup4
aiohttp
python-dateutils